added authentication and authorization

+++ RSIapi/Controllers/BoardsController.cs
@@ -1,28 +1,41 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using RSIapi.Authorization;
 using RSIapi.Context;
+using RSIapi.DTOs;
 using RSIapi.Models;
+using RSIapi.Services;
 
 namespace RSIapi.Controllers
 {
     [Route("api/boards")]
     [ApiController]
+    [Authorize]
+
     public class BoardsController : ControllerBase
     {
         private readonly ToDoItemContext _context;
+        private readonly IAuthorizationService _authorizationService;
+        private readonly IUserContextService userContextService;
 
-        public BoardsController(ToDoItemContext context)
+        public BoardsController(ToDoItemContext context, IAuthorizationService authorizationService, IUserContextService userContextService)
         {
             _context = context;
+            _authorizationService = authorizationService;
+            this.userContextService = userContextService;
         }
 
         // GET: api/Boards
         [HttpGet]
+        //[Authorize(Policy = "HasAge")] // This is the policy name from Startup.cs used for authorization by claim.
+        [Authorize(Policy = "AtLeast20")] // This is the policy name from Startup.cs used for authorization by requirement.
         public async Task<ActionResult<IEnumerable<Board>>> GetBoards()
         {
           if (_context.Boards == null)
@@ -39,10 +52,10 @@ namespace RSIapi.Controllers
         [HttpGet("{id}")]
         public async Task<ActionResult<Board>> GetBoard(int id)
         {
-          if (_context.Boards == null)
-          {
-              return NotFound();
-          }
+            if (_context.Boards == null)
+            {
+                return NotFound();
+            }
             var board = await _context.Boards.Include(b => b.ToDoItems).ThenInclude(item => item.User).FirstOrDefaultAsync(b => b.Id == id);
 
             if (board == null)
@@ -95,9 +108,18 @@ namespace RSIapi.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> PutBoard(int id, Board board)
         {
+
             if (id != board.Id)
             {
                 return BadRequest();
+            }
+
+            var authorizationResult = _authorizationService
+                .AuthorizeAsync(userContextService.User, board, new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if(!authorizationResult.Succeeded)
+            {
+                return Forbid();
             }
 
             _context.Entry(board).State = EntityState.Modified;
@@ -122,14 +144,23 @@ namespace RSIapi.Controllers
         }
 
         // POST: api/Boards
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
-        public async Task<ActionResult<Board>> PostBoard(Board board)
+        //[Authorize(Roles = "Admin,Manager")]
+        public async Task<ActionResult<Board>> PostBoard(CreateBoardDto boardDto)
         {
-          if (_context.Boards == null)
-          {
-              return Problem("Entity set 'ToDoItemContext.Boards'  is null.");
-          }
+
+            if (_context.Boards == null)
+            {
+                return Problem("Entity set 'ToDoItemContext.Boards'  is null.");
+            }
+
+            var board = new Board
+            {
+                Name = boardDto.Name,
+                Description = boardDto.Description,
+                DueTime = boardDto.DueTime,
+                CreatedById = userContextService.GetUserId
+            };
             _context.Boards.Add(board);
             await _context.SaveChangesAsync();
 
@@ -149,6 +180,14 @@ namespace RSIapi.Controllers
             if (board == null)
             {
                 return NotFound();
+            }
+
+            var authorizationResult = _authorizationService
+                .AuthorizeAsync(userContextService.User, board, new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
             }
 
             _context.Boards.Remove(board);
